kami_airlines/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
coverage==7.3.2
Django==4.2.8
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
